backend/requirements.txt
# Django et extensions
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3
django-simple-history==3.4.0
django-jazzmin==3.0.1

# Base de données
psycopg2-binary==2.9.9

# Variables d'environnement
python-dotenv==1.0.0

# Authentification et sécurité
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Utilitaires
Pillow==10.1.0
celery==5.3.4
redis==5.0.1

# Développement
django-debug-toolbar==4.2.0
django-silk==5.0.4

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Tests
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Documentation
drf-spectacular==0.26.5
